chore: move api integration tests outside of API-service
