--- services/API-service/jest.api.config.js
@@ -1,16 +1,0 @@
-/**
- * @type {import('@jest/types').Config.InitialOptions}
- */
-module.exports = {
-  moduleFileExtensions: ['js', 'ts'],
-  transform: {
-    '^.+\\.ts?$': ['ts-jest', { tsconfig: '<rootDir>/test/tsconfig.json' }],
-  },
-  rootDir: '.',
-  testMatch: ['<rootDir>/test/**/*.test.ts'],
-  coverageReporters: ['json', 'lcov'],
-  modulePathIgnorePatterns: ['<rootDir>/dist/'],
-  testTimeout: 30_000,
-  verbose: true,
-  reporters: ['default'],
-};
