+++ services/API-service/package.json
@@ -16,8 +16,6 @@
     "test": "jest --config=jest.json --detectOpenHandles --forceExit --passWithNoTests",
     "test:dev": "npm test -- --watchAll",
     "test:coverage": "npm test -- --coverage --coverageDirectory=coverage",
-    "test:api:all": "node --expose-gc node_modules/.bin/jest --config=jest.api.config.js --runInBand --detectOpenHandles --logHeapUsage",
-    "test:api:watch": "npm run test:api:all -- --watchAll",
     "typeorm": "ts-node -r tsconfig-paths/register ./node_modules/typeorm/cli.js",
     "migration:generate": "npm run typeorm migration:generate -- -d ./appdatasource.ts",
     "migration:run": "npm run typeorm migration:run -- -d ./appdatasource.ts",
@@ -61,7 +59,6 @@
     "@types/express": "^4.17.14",
     "@types/jest": "^26.0.20",
     "@types/node": "16.x",
-    "@types/supertest": "^6.0.2",
     "@typescript-eslint/eslint-plugin": "^5.48.0",
     "@typescript-eslint/parser": "^5.48.0",
     "atob": ">=2.1.0",
@@ -74,7 +71,6 @@
     "jest": "^24.9.0",
     "nodemon": "^2.0.7",
     "prettier": "^2.7.1",
-    "supertest": "^6.1.3",
     "swagger-ui-express": "^4.0.6",
     "ts-jest": "^24.3.0",
     "ts-node": "^10.9.1",
