--- services/API-service/test/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "types": ["node", "jest", "supertest"]
-  },
-  "exclude": ["../node_modules", "../dist", "../src/**/*"],
-  "include": ["**/*.ts"]
-}
