+++ tests/integration/.node-version
@@ -1,0 +1,1 @@
+v20.15.1
