+++ tests/integration/email/flash-flood/test-flash-flood-scenario.helper.ts
@@ -1,7 +1,7 @@
 import { JSDOM } from 'jsdom';
 
-import { DisasterType } from '../../../src/api/disaster/disaster-type.enum';
-import { FlashFloodsScenario } from '../../../src/scripts/enum/mock-scenario.enum';
+import { DisasterType } from '../../../../services/API-service/src/api/disaster/disaster-type.enum';
+import { FlashFloodsScenario } from '../../../../services/API-service/src/scripts/enum/mock-scenario.enum';
 import {
   getEventTitle,
   mockFlashFlood,
