+++ tests/integration/package.json
@@ -1,0 +1,43 @@
+{
+  "name": "e2e",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "lint": "eslint \"**/*.{ts,js}\"",
+    "fix": "npm run lint -- --fix",
+    "test": "node --expose-gc ./node_modules/jest/bin/jest.js --config=jest.api.config.js --runInBand --detectOpenHandles --logHeapUsage",
+    "test:watch": "npm run test:api:all -- --watchAll"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@ianvs/prettier-plugin-sort-imports": "^4.3.0",
+    "@types/express": "^4.17.14",
+    "@types/jest": "^26.0.20",
+    "@types/node": "16.x",
+    "@types/supertest": "^6.0.2",
+    "@typescript-eslint/eslint-plugin": "^5.48.0",
+    "@typescript-eslint/parser": "^5.48.0",
+    "atob": ">=2.1.0",
+    "deep-extend": ">=0.5.1",
+    "dotenv": "^16.4.5",
+    "eslint": "^8.30.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-jest": "26.x",
+    "eslint-plugin-prettier": "^3.3.1",
+    "extend": ">=3.0.2",
+    "jest": "^24.9.0",
+    "nodemon": "^2.0.7",
+    "prettier": "^2.7.1",
+    "supertest": "^6.1.3",
+    "swagger-ui-express": "^4.0.6",
+    "ts-jest": "^24.3.0",
+    "ts-node": "^10.9.1",
+    "tsc-watch": "^6.0.0",
+    "tsconfig-paths": "^3.9.0",
+    "xlsx": "^0.18.5",
+    "yargs": "^17.0.0"
+  }
+}
